Updates the behavior of the component so that the button goes away after the action is taken by the user.
