+++ src/pages/dar_application/ProgressReportApplication.tsx
@@ -15,16 +15,13 @@ import DarCloseout from 'src/pages/progress_reports/DarCloseout';
 import {CloseoutReview} from 'src/pages/progress_reports/CloseoutReview';
 import SubmitProgressReport from 'src/pages/progress_reports/SubmitProgressReport';
 import IrbDocumentUpload from 'src/pages/progress_reports/IrbDocumentUpload';
-import {Navigation, Notifications} from 'src/libs/utils';
+import {Navigation} from 'src/libs/utils';
 import {Storage} from 'src/libs/storage';
 import {DataUseAcknowledgements} from 'src/pages/dar_application/DataUseAcknowlegements';
 import {translateDataUseRestrictionsFromDataUseArray} from 'src/libs/dataUseTranslation';
 import {validatePRFormData, validationFailed} from 'src/utils/darFormUtils';
 import {FormValidationState} from 'src/pages/dar_application/FormValidationState';
 import { getApprovedElectionDatasetIds } from 'src/utils/DarUtils';
-import { DAR } from 'src/libs/ajax/DAR';
-import { User } from 'src/libs/ajax/User';
-import { AxiosError } from 'axios';
 
 type ProgressReportApplicationProps = {
   readonly dar: CombinedDataAccessRequest; // corresponds either to the parent DAR for a new application or an existing readonly progress report
@@ -175,24 +172,6 @@ export const ProgressReportApplication = ({ dar, datasets, readOnlyMode = true, 
             (user.isChairPerson && isCloseoutApproved);
     }
 
-    const onApproveReview = async () => {
-        const user = Storage.getCurrentUser();
-        const isCloseoutApproved = dar.closeoutSigningOfficialApprovedDate !== undefined;
-        try {
-            if (user.isSigningOfficial && !isCloseoutApproved) {
-                await DAR.approveCloseout(dar.referenceId);
-            } else {
-                const acknowledgement = 'dar_closeout_chair_ref_' + dar.referenceId;
-                await User.acceptAcknowledgments(acknowledgement);
-            }
-            Notifications.showSuccess({ text: 'Closeout review approved successfully.' });
-        } catch (e) {
-            const err = e as AxiosError<Record<string, string>>;
-            const message = err.response?.data.message;
-            Notifications.showError({ text: 'Error approving closeout review: ' + message });
-        }
-    }
-
     function getDatasetsInApprovedElections(datasets: Dataset[], approvedElectionDatasetIds: number[]) {
       const datasetsInApprovedElections = datasets.filter(dataset => {
         return approvedElectionDatasetIds.includes(dataset.datasetId)
@@ -296,8 +275,7 @@ export const ProgressReportApplication = ({ dar, datasets, readOnlyMode = true, 
             </div>
             {isCloseoutReview() && <div className={readOnlyMode ? 'accordion-step-container' : 'step-container'}>
                 <CloseoutReview
-                    referenceId={dar.referenceId}
-                    onApprove={onApproveReview}
+                    dar={dar}
                     onReturn={() => {
                         Navigation.console(Storage.getCurrentUser(), history);
                     }}
